+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ------
 
+## [Unreleased]
+
+### Fixed
+- Fixed `AttributeError` in `TBModel.visualize_3d()` and `Lattice.visualize_3d()` on Matplotlib >= 3.9, which removed `matplotlib.cm.get_cmap()`
+
 ## [2.0.2] - 2026-04-21
 
 ### Fixed